fix(ui-next): stop Appearance discarding the titlebar's theme choice at the next launch
`remember()` read all three axes off the DOCUMENT, so picking an accent
stored whatever `data-theme` happened to be on the root — and boot's
`applyNextDesignTheme()` puts dark there for every reader whose classic
preference resolves dark, which is the default. Pick Teal, later click
the titlebar's sun for light, restart: boot applied light and then the
accent-pick's stray `theme: "dark"` put dark back. The reader's most
recent explicit theme choice was gone, with nothing to say why.

Fixed both ways the review offered, because each alone leaves a case:

- writes are PER-AXIS now, merged onto what was STORED rather than onto
  the document, so a stored theme exists only where a reader chose one;
- and the two writers share one record. The store moves to
  `lib/appearance.ts` so the titlebar can reach it, and `Chrome`'s theme
  button calls `rememberTheme()` straight after the host's toggle — which
  reads the value the host actually put on the root rather than a mode
  this package guessed at.

Last explicit choice wins, whichever surface made it, and the tests
assert that in both orders so neither writer is privileged.
